fizzbuzz/requirements.txt
Django==1.8.6
djangorestframework==3.1.1